requirements.txt
Flask==3.0.0
requests==2.31.0
Flask-Limiter==3.5.0
cryptography==41.0.7
PyJWT==2.8.0
gunicorn==21.2.0
redis==5.0.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
sqlalchemy==2.0.23
qrcode[pil]==7.4.2
authlib==1.3.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.111.0
msal==1.26.0